[스프링 - 협업미션] 웨지(성시형) 미션 제출합니다. (#8)
* feat: 설정파일 나누기

* feat: prod.yml 파일명 수정

* feat: application.yml profiles.active local로 추가

* feat: application.yml profiles.active prod 추가

* feat: application.yml profiles.active prod 수정

* feat: 설정파일 수정

* feat: sql `` 추가

* feat: mysql ssl 설정 true

* feat: schema `` 제거

* feat: 테이블 대소문자 변경

* feat: 데이터로더 추가

* feat: mysql 설정 변경

* feat: mysql 설정 변경

* feat: schema drop 추가

* feat: schema drop if exists 삭제

* feat: application test 추가

* feat: application test 추가

* feat: API 문서 자동화 기능 추가

* feat: 지하철 API 명세 문서화

* feat: 로그인 API 명세 문서화

* feat: 회원관리 API 명세 문서화

* refactor: 테이블에 데이터 존재 시, DataLoader 미실행

* fix: 문서 스타일 변경

* feat: 노선 전체 조회 새로운 API 구현

* docs: 요구사항 작성

* feat: 기본 요금 조회 기능 추가

* refactor: 기본 요금 조회 기능 리팩토링

* refactor: 기본 요금 조회 기능 enum 리팩토링

* feat: 로그인 인터셉터 기능 추가 구현

* docs: todo 체크

* test: 테스트 버그 수정

* fix: 인터셉터 path 수정

* fix: 인터셉터 path 수정

* feat: Line 추가 운임 필드 추가 및 dao 수정

* feat: Line 추가 운임 관련 service, controller 수정

* feat: 추가 운임 계산 로직 구현

* style: 공백 라인 제거

* refactor: 서비스에 Transactional 추가

* feat: 노선 중복 색 입력시 예외처리

* feat: 노선 중복 색 입력시 예외처리

* feat: 멤버 이메일 중복 시 예외처리

* feat: 연령별 요금 제도 추가 구현

* feat: 노선 삭제시 해당 구간도 함께 삭제

* test: 테스트 코드 중복 제거

* feat: 없는 역 아이디 조회시 예외처리

* refactor: 코드 리팩토링

* refactor: API 명세 문서에 토큰 헤더 추가

* refactor: 역 조회시 아이디순으로 정렬

* feat : Fare VO 추가

* refactor : Exception 관련 클래스 패키지 수정

* refactor : 적절한 클래스명으로 수정

ExceptionController.java -> ControllerExceptionAdvice.java

* refactor : 적절한 상수명과 인자명으로 수정

1. km -> kmUnit

* refactor : 나이에 따른 할인전략인 AgeFarePolicy 생성

* refactor : 게스트 객체를 삭제하고 로그인 멤버가 나이값이 없으면(게스트인 경우) 기본 요금을 반환하는 방식으로 리팩토링

* refactor : 사용하지 않는 import 제거

* test : Fare 도메인 테스트 추가

* test : guest에 대한 요금 계산 테스트 추가

* test & refactor : 요금 계산 공식 클래스 테스트와 상수 추출

* refactor : test data loader의 데이터를 호출시 생성하는 방식으로 변경

* refactor : guest 판단 로직을 Role enum을 통해 판단하도록 수정

+++ src/main/java/wooteco/subway/station/application/StationService.java
@@ -1,6 +1,8 @@
 package wooteco.subway.station.application;
 
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+import wooteco.subway.line.dao.SectionDao;
 import wooteco.subway.station.dao.StationDao;
 import wooteco.subway.station.domain.Station;
 import wooteco.subway.station.dto.StationRequest;
@@ -10,19 +12,26 @@ import java.util.List;
 import java.util.stream.Collectors;
 
 @Service
+@Transactional
 public class StationService {
     private StationDao stationDao;
+    private final SectionDao sectionDao;
 
-    public StationService(StationDao stationDao) {
+    public StationService(StationDao stationDao, SectionDao sectionDao) {
         this.stationDao = stationDao;
+        this.sectionDao = sectionDao;
     }
 
+    @Transactional
     public StationResponse saveStation(StationRequest stationRequest) {
         Station station = stationDao.insert(stationRequest.toStation());
         return StationResponse.of(station);
     }
 
     public Station findStationById(Long id) {
+        if (stationDao.notExistId(id)) {
+            throw new NotFoundStationException();
+        }
         return stationDao.findById(id);
     }
 
@@ -34,7 +43,16 @@ public class StationService {
                 .collect(Collectors.toList());
     }
 
+    @Transactional
     public void deleteStationById(Long id) {
+        if (stationDao.notExistId(id)) {
+            throw new NotFoundStationException();
+        }
+
+        if (sectionDao.existSectionByStationId(id)) {
+            throw new IncludedStationException();
+        }
+
         stationDao.deleteById(id);
     }
 }
